Consolidate gff3 adapter featureData routine, avoid number-ifying null phase
